meson: use C variable names for IR
Clang defaults to drop C/C++ variable names when translating to IR and replace
 the names with numbers. This can be controlled with the argument
"-discard-value-names" in the Clang frontend mode (-cc1) and is inserted
automatically when calling a release build of clang. This option, however, is
exported via "-fno-discard-value-names" to the complete compiler but only
within Clang-7.

Variable names in the IR improve readability.