logging: drop dynamic formatting
This does not work in this way.
Problem: The stepmanager setup all step loggers when initializing
itself. If logging.basicConfig is called after that, all step loggers
are getting the warning log level. If we want to set the format
depending of the step name length, we have to set it after the
stepmanager initialization.
So drop dynamic initialization in favor of the correct log level (for
now).