steps: put template functions in anonymous namespace
The template functions to handle all graph actions are defined in all
steps that handle the graph with nearly similar arguments and maybe the
same function name.

With Clang this leads to a symbol conflict after compilation. That
has resulted in the execution of do_graph_stuff function from the
cdummy step instead of the do_graph_stuff function of the
value_analysis_core step.