os.py: use getattr to retrieve the syscall
This is a much cleaner code, since it does not depend on Python
implementation internals. However, this has not worked in the first
place, because staticmethod is creating a descriptor that is called
transparently by getattr, so the function attribute was invisible.
Attaching the attribute to the original function fixes this.

See also https://docs.python.org/3/howto/descriptor.html