appl: FreeRTOS: queue sending to higher and lower prio
expected behaviour:
sender sends ttttuuuuvvvvwwww
it is expected to be interrupted by higher prio receiver printing each
character in uppercase.
Thus, we expect the following output:
TtTtTtUuUuUuVvVvVvWwWwWw

Nect, we send the same sequence ttttuuuuvvvvwwww to a lower prio
receiver task.
here, we expect the sender to continue after sending. The queue has a
capacity of three elements, resulting in blocking the sender after
three t's, then alternating with the receiver and a fan out of the
receiver.
The second receiver prints in ROT13 to be distinguishable from the
first one.
tttugugugvhvhvhwiwiwijjj