FlatAnalysis: fix call path recognition
When calling a function B twice in a function A and calling then a
function C within the function B, the call path is recognized in a wrong
way. That is because they have the same cfglink (aka calling ABB). The
code before this patch uses a flat mapping to map the cfglink to the
call-graph node. Therefore a calling ABB was always mapped to exactly
the same call-graph node. With this patch the old call path is respected,
so that a calling ABB can be mapped to multiple call-graph nodes (the
node in the correct subtree).

Fixes T155.
Reported-by: Björn Fiedler <fiedler@sra.uni-hannover.de>