avoid use of root python logger and log level
The Python logging framework is designed in such a way that the log
level of the root logger is the maximum level. Even is child logger has
more detailed levels, they aren't logging anything.

To circumvent this, we introduce our own level that is stored globally
and set the root logger always to debug.

The downside is that we aren't allowed to use global logging anymore.

Therefore always use:
```
from ara.util import get_logger
_log = get_logger("MySubsystem")
_log.info("Hi there!")
```