steps/icfg: make indirect pointer analysis sound
Previously, we accepted all functions with the same signature for an
unresolved indirect pointer. However, this is not sound, since the LLVM
linker merges types with the same bitsize [1], resulting in missing
candidate functions.

Now, we only check for the same amount of arguments. This is less
precise, but sound.

[1] http://lists.llvm.org/pipermail/llvm-dev/2020-May/141900.html