ValueAnalysis: don't abort too early
The value analysis does a call path dependent lookup, thus retrieving at
maximum one value per call path. However, the SVFG has function local
and interprocedural value flows, so the value analysis has to store
whether it has already found a value on this particular function level.

Up to this commit there was a bug in the bookkeeping, since multiple
functions with the same depth in the call path are treated equally.