Transition from ValueAnalysis to inline ValueAnalyzer
Before this commit, the value analysis was done as preprocessing step
for all syscalls at once. Now it is done directly, when needed.
Typically, this is as part of the OS model.

As part of the transition the syscall decorator is replaced with
syscall2 which also assigns the argument names and triggers the value
analysis.

What is further done is an assignment of system objects within the SVFG
i.e. in an extra map. This map is currently not globally stored but as
side data of the value analyzer.

Also, the traversal method of the SVFG changed. Before, it was more less
a depth first search. Now it is more less a breadth first search.