CFG: extend file and line to lists
Now all files and all lines of all instructions in the basic block/ABB
are stored.

Not that a list of files is also necessary since functions can be
declared as "forceinlined". In this case, even an unoptimized Clang
run will inline the function so the basic block contains instructions
from different files.