meson: fix shared module linking
Python binary modules are shared modules. However, linking them together
is not supported. Therefore, they need to be loaded one after another in
the Python interpreter so it can resolve the symbols from the second
module also with the one from the first. This is achieved with an extra
flag to dlopen (the mechanism behind Python module loading).

To proper link the module, we also have to tell the linker to allow
undefined symbols.